resource manifest write, page post processor, lot of good stuff

+++ skydump.py
@@ -4,13 +4,15 @@ import os
 import logging
 import shutil
 import mimetypes
+import json
 from dataclasses import dataclass, asdict
 from typing import List, Set
 
 from bs4 import BeautifulSoup, PageElement, Tag
 
-from models.link import Link, LinkType
+from models.link import Link
 from models.page import Page
+from models.resource import Resource
 
 
 REG_DOMAIN = re.compile(r"([a-zA-Z]*://[^\/]+)", re.I)
@@ -21,6 +23,7 @@ REG_RESOURCE_OLD = re.compile(r'"((?:[a-zA-Z]*://)?[^ "]+)"', re.I)
 REG_RESOURCE = re.compile(r'((?:[a-zA-Z]*:\/)?\/.+)', re.I)
 REG_URL_NO_PROTOCOL = re.compile(r"^([a-zA-Z]+://)?([^\/]*)(\/[^?]*)?(\?.*)?", re.I)
 #(.+)
+
 
 def parse_page(url,
                allow_crawl_conditions: List[re.Pattern] = list(),
@@ -34,11 +37,11 @@ def parse_page(url,
     forbid_crawl_conditions: list of regexes that must NOT match to allow the link to be entered
     """
     
-    domain_reg = REG_DOMAIN.search(url)
+    domain_reg = REG_URL_NO_PROTOCOL.search(url)
     if not domain_reg:
-        raise Exception(f"Can't parse domain in url: {url}")
+        raise Exception(f"Can't parse url: {url}")
 
-    curr_page = Page(remote_url=url, domain=domain_reg.group(1))
+    curr_page = Page(remote_url=url, domain=domain_reg.group(2), protocol=domain_reg.group(1))
 
     logging.info(f"Requesting page {url}")
 
@@ -62,14 +65,24 @@ def parse_page(url,
 
                 logging.info(f"Found link {link_str}")
 
-                new_link.type = LinkType.PAGE
+                url_reg = REG_URL_NO_PROTOCOL.search(link_str)
+                if not url_reg:
+                    raise Exception(f"Can't parse resource url: {link_str}")
+                
+                new_link.resource = Page()
                 new_link.original_url = link_str
 
+                new_link.resource.remote_url = link_str
+                new_link.resource.protocol=url_reg.group(1)
+                new_link.resource.domain=url_reg.group(2)
+
                 if link_str.startswith("/"):
-                    new_link.remote_url = curr_page.domain + link_str
+                    new_link.resource.remote_url = curr_page.protocol + curr_page.domain + link_str
+                    new_link.resource.protocol = curr_page.protocol
+                    new_link.resource.domain = curr_page.domain
                 
                 elif all(map(lambda r: r.search(link_str) is not None, allow_crawl_conditions)) and all(map(lambda r: r.search(link_str) is None, forbid_crawl_conditions)):
-                    new_link.remote_url = link_str
+                    new_link.resource.remote_url = link_str
                     
                 else:
                     logging.info(f"Ignored link {link_str}")
@@ -85,11 +98,18 @@ def parse_page(url,
                         
                         logging.info(f"Found resource {link_str}")
 
-                        new_link.type = LinkType.RESOURCE
-                        new_link.original_url = link_str
-                        new_link.remote_url = link_str
+                        url_reg = REG_URL_NO_PROTOCOL.search(link_str)
+                        if not url_reg:
+                            raise Exception(f"Can't parse resource url: {link_str}")
 
-            if new_link.remote_url:
+                        new_link.resource = Resource()
+                        new_link.original_url = link_str
+
+                        new_link.resource.remote_url = link_str
+                        new_link.resource.protocol=url_reg.group(1)
+                        new_link.resource.domain=url_reg.group(2)
+
+            if new_link.resource and new_link.resource.remote_url:
                 curr_page.links.append(new_link)
     
         return curr_page
@@ -109,14 +129,10 @@ def parse_url(url, force_extension=None):
     add_html_extension: always add a .html to the name if the resource is not already a .htm/.html
     """
 
-    print(f"{url=}")
-
     res = REG_URL_NO_PROTOCOL.search(url)
     if not res:
         logging.warning(f"Can't download page {url}: non conform URL ")
         return None
-
-    print(f"{res.groups()=}")
 
     parameters = res.group(4)
 
@@ -150,10 +166,12 @@ def parse_url(url, force_extension=None):
         # TODO: somehow vérifier le ContentType avant de faire ça
         if force_extension is not None:
             page_name += force_extension
+    
+    page_path = [res.group(2)] + page_path
 
-    print(f"{page_path=}, {page_name=}")
-    print(f"{res.group(2)}/{'/'.join(page_path)}/{page_name}")
-    print()
+    #print(f"{page_path=}, {page_name=}")
+    #print(f"{res.group(2)}/{'/'.join(page_path)}/{page_name}")
+    #print()
 
     return page_path, page_name
 
@@ -167,14 +185,14 @@ def find_mimetype(header_content_type: str) -> str:
         return None
 
 
-def download_page(url, destination_path):
+def download(url, destination_path, overwrite=True):
     logging.info(f"Downloading {url} to {destination_path}")
 
     r = requests.get(url)
-    page_content = r.content
+    rsc_content = r.content
 
     if r.status_code != 200:
-        logging.error(f"Error code {r.status_code} while getting page {url}")
+        logging.error(f"Error code {r.status_code} while getting resource {url}.")
         return None
     
     content_type = find_mimetype(r.headers["Content-Type"])
@@ -182,17 +200,23 @@ def download_page(url, destination_path):
 
     if extension:
         destination_path = os.path.splitext(destination_path)[0] + extension
-        
-    with open(destination_path, "wb") as fp:
-        if r.apparent_encoding:
-            fp.write(page_content.decode(r.apparent_encoding).encode("utf-8"))
-        else:
-            fp.write(page_content)
+    
+    if overwrite is False and os.path.exists(destination_path):
+        logging.info(f"Resource {destination_path} already exist. Skipping download!")
+    else:
+        with open(destination_path, "wb") as fp:
+            if r.apparent_encoding:
+                fp.write(rsc_content.decode(r.apparent_encoding).encode("utf-8"))
+            else:
+                fp.write(rsc_content)
     
     return destination_path, content_type
 
 
-def remap_html_page(origin_local_url, original_url, local_url):
+def remap_html_page(origin_local_url, original_url, local_url, relative=True):
+    if relative:
+        local_url = (len(origin_local_url.split("/"))-1) * "../" + local_url
+
     logging.info(f"Remapping file {origin_local_url} by replacing {original_url} with {local_url}")
 
     local_file_content = None
@@ -203,48 +227,85 @@ def remap_html_page(origin_local_url, original_url, local_url):
         fp.write(local_file_content.replace(f'"{original_url}"', local_url).encode("utf-8"))
 
 
-post_process_link_resource_by_extension = {
-    "css": [
+page_post_processors = [
+    lambda p, l: remap_html_page(p.local_url, l.resource.remote_url, l.resource.local_url, relative=True)
+]
 
+asset_post_processors = {
+    "text/css": [
+        
     ]
 }
 
-post_process_page_by_extension = {
-    "html": [
-        lambda p, l: remap_html_page(p.local_url, l.remote_url, l.local_url)
-    ],
-}
 
+def get_resource_local_url(remote_url):
+    rsc_path, rsc_name = parse_url(remote_url)
+    return "/".join(rsc_path) + "/" + rsc_name
+
+
+def retrieve_resource(remote_url, destination_path, overwrite=True):
+    destination_dir = os.path.dirname(destination_path)
+
+    if destination_dir != "" and not os.path.exists(destination_dir):
+        os.makedirs(destination_dir)
+
+    downloaded_file, content_type = download(remote_url, destination_path, overwrite)
+    return downloaded_file, content_type
+
+
+def write_resource_manifest(rsc: Resource, path: str = None):
+    if path is None:
+        path = os.path.splitext(rsc.local_url)[0] + ".json"
+    
+    logging.info(f"Writing resource {rsc.remote_url} manifest in {path}")
+
+    with open(path, "w") as fp:
+        fp.write(json.dumps(asdict(rsc), indent=4))
+
+
+# TODO: POUR REGLER PROBLMEE DE DEVOIR DL LE FICHIER POUR AVOIR SON MIME
+# ON LE DL LA PREMIERE FOIS ET ON CREE UN JSON A COTE AVEC LE NOM DU TRUC BASIQUE + JSON QUI MAPPE
+# LE vRAI FICHIER AVEC L'URL DE BASE ETC
 def execute(url,
             allow_crawl_conditions: List[re.Pattern] = list(),
             forbid_crawl_conditions: List[re.Pattern] = list()):
 
     # Retrieve page and its allowed linked pages & resources
     page = parse_page(url, allow_crawl_conditions, forbid_crawl_conditions)
+    
+    page.local_url = get_resource_local_url(page.remote_url)
+    downloaded_file, content_type = retrieve_resource(page.remote_url, page.local_url)
+    page.content_type = content_type
+
+    write_resource_manifest(page)
+
+    # Backing up original page
+    shutil.copyfile(downloaded_file, downloaded_file + ".orig")
 
     for l in page.links:
-        page_path, page_name = parse_url(l.remote_url)
+        destination_path = get_resource_local_url(l.resource.remote_url)
+        l.resource.local_url = destination_path
 
-        path = "/".join(page_path)
-
-        if not os.path.exists(path):
-            os.makedirs(path)
+        downloaded_file, content_type = retrieve_resource(l.resource.remote_url,
+                                                          destination_path,
+                                                          overwrite=False)
         
-        destination_path = path + "/" + page_name
-        print(f"{destination_path=}")
-        downloaded_file, content_type = download_page(l.remote_url, destination_path)
+        # Updating the local_url field with the real local url of thed ownloaded file
+        # (to integrate corrected extension detected from the mimetype)
+        l.resource.local_url = downloaded_file
+        l.resource.content_type = content_type
 
-        l.local_url = downloaded_file
-        l.content_type = content_type
+        write_resource_manifest(l.resource)
 
-        # Backing up original page
-        if l.type == LinkType.PAGE:
-            shutil.copyfile(downloaded_file, downloaded_file + ".orig")
-    
     # Run post-process operations
-    #TODO: A TESTER
     for link in page.links:
-        post_process_page_by_extension["html"](page, link)
+        for fn in page_post_processors:
+            fn(page, link)
+
+    for link in page.links:
+        fn_list = asset_post_processors.get(link.resource.content_type, [])
+        for fn in fn_list:
+            fn(page, link)
 
     # # Remap page links & resources to local equivalents
     # origin_local_page_path, origin_local_page_name = parse_url(url)
